+++ src/test/java/de/codescape/jira/plugins/scrumpoker/rest/entities/CardEntityTest.java
@@ -1,7 +1,9 @@
 package de.codescape.jira.plugins.scrumpoker.rest.entities;
 
+import org.codehaus.jackson.map.ObjectMapper;
 import org.junit.Test;
 
+import static org.hamcrest.Matchers.containsString;
 import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.is;
 import static org.junit.Assert.assertThat;
@@ -15,6 +17,13 @@ public class CardEntityTest {
         CardEntity cardEntity = new CardEntity(CARD_VALUE, true);
         assertThat(cardEntity.getValue(), is(equalTo(CARD_VALUE)));
         assertThat(cardEntity.isSelected(), is(equalTo(true)));
+    }
+
+    @Test
+    public void jsonRepresentationContainsAllFields() throws Exception {
+        String json = new ObjectMapper().writeValueAsString(new CardEntity(CARD_VALUE, false));
+        assertThat(json, containsString("value"));
+        assertThat(json, containsString("selected"));
     }
 
 }