+++ src/test/java/de/codescape/jira/plugins/scrumpoker/rest/entities/SessionEntityTest.java
@@ -1,7 +1,9 @@
 package de.codescape.jira.plugins.scrumpoker.rest.entities;
 
+import org.codehaus.jackson.map.ObjectMapper;
 import org.junit.Test;
 
+import static org.hamcrest.Matchers.containsString;
 import static org.hamcrest.Matchers.is;
 import static org.junit.Assert.assertThat;
 
@@ -15,6 +17,20 @@ public class SessionEntityTest {
     @Test
     public void shouldSignalNoExistenceOfConfirmedVoteIfNoConfirmedVoteIsSet() {
         assertThat(new SessionEntity().withConfirmedVote(null).isConfirmedVoteExists(), is(false));
+    }
+
+    @Test
+    public void jsonRepresentationsContainsAllFields() throws Exception {
+        String json = new ObjectMapper().writeValueAsString(new SessionEntity());
+        assertThat(json, containsString("issueKey"));
+        assertThat(json, containsString("cards"));
+        assertThat(json, containsString("confirmedVote"));
+        assertThat(json, containsString("visible"));
+        assertThat(json, containsString("boundedVotes"));
+        assertThat(json, containsString("agreementReached"));
+        assertThat(json, containsString("votes"));
+        assertThat(json, containsString("allowReset"));
+        assertThat(json, containsString("allowReveal"));
     }
 
 }