+++ src/test/java/de/codescape/jira/plugins/scrumpoker/rest/entities/VoteEntityTest.java
@@ -1,9 +1,9 @@
 package de.codescape.jira.plugins.scrumpoker.rest.entities;
 
+import org.codehaus.jackson.map.ObjectMapper;
 import org.junit.Test;
 
-import static org.hamcrest.Matchers.equalTo;
-import static org.hamcrest.Matchers.is;
+import static org.hamcrest.Matchers.*;
 import static org.junit.Assert.assertThat;
 
 public class VoteEntityTest {
@@ -17,6 +17,14 @@ public class VoteEntityTest {
         assertThat(voteEntity.getUser(), is(equalTo(USER)));
         assertThat(voteEntity.getVote(), is(equalTo(VOTE)));
         assertThat(voteEntity.isNeedToTalk(), is(equalTo(true)));
+    }
+
+    @Test
+    public void jsonRepresentationContainsAllFields() throws Exception {
+        String json = new ObjectMapper().writeValueAsString(new VoteEntity(USER, VOTE, false));
+        assertThat(json, containsString("user"));
+        assertThat(json, containsString("vote"));
+        assertThat(json, containsString("needToTalk"));
     }
 
 }